Add endpoint for saving user profile image to S3
This endpoint pulls down the uploaded image from a
temporary S3 bucket that deletes all objects after 1 day.

Process the image and then save it to a permanent S3
bucket.

Amazon S3 allows uploads direct from the browser so
Node.js doesn't have to handle the bandwidth of uploading
files.

Uploading to the S3 temporary bucket happens on the client.