Modify endpoint for avatar upload--upload to NodeJS
The previous version was overkill; uploading to a temporary
bucket client-side, then pulling it down, processing it, then
uploading it to a permanent bucket from Node.

Now client uploads avatar directly to Node which processes it
and uploads it to a permanent bucket.

If dealing with larger files maybe the former would be a wise
choice. But for avatars the latter will do fine.