Add unreadComments to User schema
Now users will have an unreadComments array. When someone
comments on their poem, the comment's id will be pushed into
the array. The client can then display a notification if the
array is not empty. Once the client requests the unread
comments' poem, the id will be removed from the array.